+++ src/basemath/prime.c
@@ -1230,7 +1230,7 @@ primes_interval_zv(ulong a, ulong b)
   if (!a) return primes_upto_zv(b);
   if (b < a) return cgetg(1, t_VECSMALL);
   d = b - a;
-  if (d > 100000UL) d = primepi_upper_bound(b) - primepi_lower_bound(a);
+  if (d > 100000UL) d = (ulong)(primepi_upper_bound(b)-primepi_lower_bound(a));
   return primes_interval_i(a, b, d);
 }
 GEN
@@ -1238,7 +1238,7 @@ primes_upto_zv(ulong b)
 {
   ulong d;
   if (b < 2) return cgetg(1, t_VECSMALL);
-  d = (b > 100000UL)? primepi_upper_bound(b): b;
+  d = (b > 100000UL)? (ulong)primepi_upper_bound(b): b;
   return primes_interval_i(2, b, d);
 }
